removed folders
unnecessary folders removed
