--- frontend/.babelrc
@@ -1,12 +1,0 @@
-{
-  "presets": [
-    "@babel/preset-env",
-    "@babel/preset-react"
-  ],
-  "plugins": [
-    ["@babel/plugin-proposal-decorators", { "legacy": true }],
-    "transform-class-properties",
-    "transform-react-constant-elements",
-    "transform-react-inline-elements"
-  ]
-}
