--- frontend/.editorconfig
@@ -1,19 +1,0 @@
-# EditorConfig helps developers define and maintain consistent
-# coding styles between different editors and IDEs
-# editorconfig.org
-
-root = true
-
-[*]
-end_of_line = lf
-charset = utf-8
-trim_trailing_whitespace = true
-insert_final_newline = true
-indent_style = space
-indent_size = 2
-
-[*.{hbs,html}]
-insert_final_newline = false
-
-[*.{diff,md}]
-trim_trailing_whitespace = false
