--- frontend/.eslintrc
@@ -1,3 +1,0 @@
-{
-  "extends": "standard"
-}
