--- frontend/config/setupTest.js
@@ -1,1 +1,0 @@
-module.exports = {}
