--- frontend/package.json
@@ -1,85 +1,0 @@
-{
-  "name": "creditaschallenge",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "node server.js",
-    "test": "jest"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://gitlab.com/rafaellucio/creditaschallenge.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://gitlab.com/rafaellucio/creditaschallenge/issues"
-  },
-  "homepage": "https://gitlab.com/rafaellucio/creditaschallenge#readme",
-  "devDependencies": {
-    "@babel/cli": "^7.2.3",
-    "@babel/core": "^7.3.3",
-    "@babel/plugin-proposal-decorators": "^7.3.0",
-    "@babel/preset-env": "^7.3.1",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/preset-stage-0": "^7.0.0",
-    "@types/jest": "^23.3.2",
-    "babel-jest": "^24.1.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-react-constant-elements": "^6.23.0",
-    "babel-plugin-transform-react-inline-elements": "^6.22.0",
-    "css-loader": "^1.0.0",
-    "dom-testing-library": "^3.5.1",
-    "html-webpack-plugin": "^3.2.0",
-    "jest": "^24.1.0",
-    "jest-dom": "^2.1.1",
-    "mini-css-extract-plugin": "^0.4.2",
-    "regenerator-runtime": "^0.12.1",
-    "standard": "^12.0.1",
-    "style-loader": "^0.23.0",
-    "webpack": "^4.18.0",
-    "webpack-cli": "^3.1.0",
-    "webpack-dev-server": ">=3.1.11"
-  },
-  "jest": {
-    "browser": true,
-    "verbose": true,
-    "clearMocks": true,
-    "collectCoverage": true,
-    "testEnvironment": "jest-environment-jsdom",
-    "coverageThreshold": {
-      "global": {
-        "branches": 80,
-        "functions": 80,
-        "lines": 80,
-        "statements": -10
-      }
-    },
-    "testMatch": [
-      "**/__tests__/**/*.js?(x)",
-      "**/?(*.)+(spec|test).js?(x)"
-    ],
-    "coveragePathIgnorePatterns": [
-      "/node_modules/"
-    ],
-    "collectCoverageFrom": [
-      "src/**/*.{js,jsx}"
-    ],
-    "testPathIgnorePatterns": [
-      "/node_modules/"
-    ],
-    "transformIgnorePatterns": [
-      "/node_modules/"
-    ],
-    "setupTestFrameworkScriptFile": "<rootDir>/config/setupTest.js",
-    "transform": {
-      "^.+\\.js$": "babel-jest"
-    },
-    "moduleNameMapper": {
-      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
-      "\\.(css|scss)$": "<rootDir>/__mocks__/styleMock.js"
-    }
-  }
-}
