--- frontend/public/index.html
@@ -1,104 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Challenge Creditas</title>
-  </head>
-  <body>
-    <div class="container">
-      <header class="header">
-        <div class="header__container">
-          <h1 class="header__title">
-            <a href="#" title="Creditas">
-              <img class="logo" src="https://staging.creditas.com.br/static/images/logo-creditas-white-3cd22a2808.svg" alt="Creditas">
-            </a>
-          </h1>
-          <a class="help" href="#ajuda">Ajuda</a>
-        </div>
-      </header>
-      <main class="main">
-        <h1 class="main__title">Realize uma simulação de crédito utilizando seu bem como garantia.</h1>
-        <section class="section__container">
-          <form class="form" name="form">
-            <div class="form__fields">
-              <div class="field-group">
-                <div class="field">
-                  <label for="parcelas">Número de parcelas</label>
-                  <select name="parcelas" id="parcelas" required>
-                    <option value="24">24</option>
-                    <option value="36">36</option>
-                    <option value="48">48</option>
-                  </select>
-                </div>
-                <div class="field">
-                  <label for="garantia">Garantia</label>
-                  <select name="garantia" id="garantia" required>
-                    <option value="veiculo">Veículo</option>
-                    <option value="imovel">Imóvel</option>
-                  </select>
-                </div>
-              </div>
-              <div class="valor-garantia">
-                <div class="field-group">
-                  <div class="field">
-                    <label for="valor-garantia">Valor da Garantia</label>
-                    <input type="text" required name="valor-garantia" id="valor-garantia" value="14400" />
-                  </div>
-                  <div class="field">
-                    <div class="range">
-                      <input type="range" name="valor-garantia-range" id="valor-garantia-range" min="0" max="100" value="20" step="10">
-                      <div class="range__values">
-                        <span>12.000</span>
-                        <span>24.000</span>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-              <div class="emprestimo">
-                <div class="field-group">
-                  <div class="field">
-                    <label for="valor-emprestimo">Valor do Empréstimo</label>
-                    <input type="text" required name="valor-emprestimo" id="valor-emprestimo" value="57000">
-                  </div>
-                  <div class="field">
-                    <div class="range">
-                      <input type="range" name="valor-emprestimo-range" id="valor-emprestimo-range" min="0" max="100" value="90" step="10">
-                      <div class="range__values">
-                        <span>30.000</span>
-                        <span>60.000</span>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div class="form__result">
-              <div class="quota__container">
-                <h4>Valor da Parcela</h4>
-                <div class="quota">
-                  <strong>R$</strong>
-                  <span>465,00</span>
-                </div>
-              </div>
-              <div class="amount_container">
-                <h4>Total a pagar</h4>
-                <p>R$ 11.112,00</p>
-              </div>
-              <div class="tax__container">
-                <h4>Taxa de juros (mês)</h4>
-                <p>111,12%</p>
-              </div>
-              <button class="button">
-                Solicitar
-              </button>
-            </div>
-          </form>
-        </section>
-        <footer class="footer">*Lorem ipsum dolor sit amet consectetur adipisicing elit. Facere quas non harum dolor eligendi id, ab corrupti blanditiis suscipit ex odit error alias minus. Enim dolores eum officiis quae rem!*</footer>
-      </main>
-    </div>
-  </body>
-</html>
